* test/lib/envutil.rb: move enutil's assertions under Test::Unit::Assertion. * test/lib/test/unit/assertions.rb: ditto.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@53089 b2dd03c8-39d4-4d8f-98ff-823fe69b080e
